Added color management page

+++ app/services/update_colors.rb
@@ -2,11 +2,13 @@ class UpdateColors
   def call(params)
     params[:colors].each do |_, v|
       c = Color.find(v[:id])
-      c.tags.each do |t|
-        t.image.destroy
-        ImageWorker.perform_async(t.id)
+      if c.hex != v[:hex]
+        c.update(hex: v[:hex])
+        c.tags.each do |t|
+          t.image.destroy
+          ImageWorker.perform_async(t.id)
+        end
       end
-      c.update(name: v[:name], hex: v[:hex], complete: true)
     end
   rescue
     [false, nil]
